requirements.txt
# Core dependencies
pydantic==2.12.3
pydantic-settings==2.7.0
pydantic_core==2.41.4
python-dotenv==1.2.1

# AI/LLM
ollama==0.6.0

# Browser automation
playwright==1.55.0
pyee==13.0.0

# Document handling
python-docx==1.1.2
lxml==6.0.2

# Web scraping
beautifulsoup4==4.14.2
soupsieve==2.8
requests==2.32.5

# HTTP clients
httpx==0.28.1
httpcore==1.0.9
urllib3==2.5.0
certifi==2025.10.5

# Type annotations
typing_extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0

# Async support
anyio==4.11.0
greenlet==3.2.4
sniffio==1.3.1

# Utilities
tqdm==4.67.1
PyYAML==6.0.3
charset-normalizer==3.4.4
idna==3.11

# Testing (dev dependencies)
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.25.2

# Google APIs (optional)
google-api-python-client==2.185.0
google-auth==2.42.1
google-generativeai==0.8.5

# Legacy dependencies (can be removed if not used)
ollama-python==0.1.2
openai==2.6.1
